Bugfix baseline_critic: ensure observations are equal
ISSUE:
-----
Observations are described with agents and their relative positions.

FIX:
---
Observations are swapped to ensure that all agents get the same
representation. Change from relative position representation to
absolute. As opposed to previous fix that did not work on partial
environments this fix preserves the partial observability for the
assigned agent.